+++ go/src/github.com/bitrise-core/bitrise-init/scanners/ios/ios_test.go
@@ -1,8 +1,11 @@
 package ios
 
 import (
+	"os"
+	"path"
 	"testing"
 
+	"github.com/bitrise-io/go-utils/pathutil"
 	"github.com/stretchr/testify/require"
 )
 
@@ -209,7 +212,9 @@ func TestIsRelevantProject(t *testing.T) {
 		}
 
 		for _, file := range fileList {
-			require.Equal(t, false, isRelevantProject(file))
+			is, err := isRelevantProject(file, true)
+			require.NoError(t, err)
+			require.Equal(t, false, is)
 		}
 	}
 
@@ -220,8 +225,28 @@ func TestIsRelevantProject(t *testing.T) {
 		}
 
 		for _, file := range fileList {
-			require.Equal(t, true, isRelevantProject(file))
+			is, err := isRelevantProject(file, true)
+			require.NoError(t, err)
+			require.Equal(t, true, is)
 		}
+	}
+
+	t.Log("symlink is not valid")
+	{
+		tmpDir, err := pathutil.NormalizedOSTempDirPath("test")
+		require.NoError(t, err)
+
+		pth := path.Join(tmpDir, "SampleAppWithCocoapods.xcodeproj")
+		require.NoError(t, os.MkdirAll(pth, 0777))
+		is, err := isRelevantProject(pth, false)
+		require.NoError(t, err)
+		require.Equal(t, true, is)
+
+		symlinkPth := path.Join(tmpDir, "symlink-SampleAppWithCocoapods.xcodeproj")
+		require.NoError(t, os.Symlink(pth, symlinkPth))
+		is, err = isRelevantProject(symlinkPth, false)
+		require.NoError(t, err)
+		require.Equal(t, false, is)
 	}
 }
 
@@ -237,7 +262,8 @@ func TestFilterXcodeprojectFiles(t *testing.T) {
 			"/Users/bitrise/sample-apps-ios-cocoapods/SampleAppWithCocoapods.xcodeproj",
 		}
 
-		files := filterXcodeprojectFiles(fileList)
+		files, err := filterXcodeprojectFiles(fileList, true)
+		require.NoError(t, err)
 		require.Equal(t, 1, len(files))
 		require.Equal(t, "/Users/bitrise/sample-apps-ios-cocoapods/SampleAppWithCocoapods.xcodeproj", files[0])
 	}
@@ -253,7 +279,8 @@ func TestFilterXcodeprojectFiles(t *testing.T) {
 			"SampleAppWithCocoapods.xcodeproj",
 		}
 
-		files := filterXcodeprojectFiles(fileList)
+		files, err := filterXcodeprojectFiles(fileList, true)
+		require.NoError(t, err)
 		require.Equal(t, 1, len(files))
 		require.Equal(t, "SampleAppWithCocoapods.xcodeproj", files[0])
 	}
@@ -265,7 +292,8 @@ func TestFilterXcodeprojectFiles(t *testing.T) {
 			"/Users/bitrise/sample-apps-ios-cocoapods/SampleAppWithCocoapods.xcworkspace",
 		}
 
-		files := filterXcodeprojectFiles(fileList)
+		files, err := filterXcodeprojectFiles(fileList, true)
+		require.NoError(t, err)
 		require.Equal(t, 2, len(files))
 
 		require.Equal(t, "/Users/bitrise/sample-apps-ios-cocoapods/SampleAppWithCocoapods.xcworkspace", files[0])
@@ -280,7 +308,8 @@ func TestFilterXcodeprojectFiles(t *testing.T) {
 			"build.gradle",
 		}
 
-		files := filterXcodeprojectFiles(fileList)
+		files, err := filterXcodeprojectFiles(fileList, true)
+		require.NoError(t, err)
 		require.Equal(t, 0, len(files))
 	}
 }